Refactorzito & Adição de Valores Ficheiro strings.xml
Refactorzito no Nomes dos Fragmentos (Home , Biography & Mapa)
Adição de Valores ao ficheiro strings.xml
