+++ app/src/main/java/com/example/raullino/ui/biography/BiographyViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.raullino.ui.notifications
+package com.example.raullino.ui.biography
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class BiographyViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "Biography Fragment"
     }
     val text: LiveData<String> = _text
 }
