+++ app/src/main/java/com/example/raullino/ui/home/HomeViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.ViewModel
 class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "Home Fragment"
     }
     val text: LiveData<String> = _text
 }
