+++ app/src/main/java/com/example/raullino/ui/map/MapViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.raullino.ui.dashboard
+package com.example.raullino.ui.map
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class MapViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "Map Fragment"
     }
     val text: LiveData<String> = _text
 }
