+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,6 @@
     <string name="title_home">Home</string>
     <string name="title_dashboard">Dashboard</string>
     <string name="title_notifications">Notifications</string>
+    <string name="mapa">Mapa</string>
+    <string name="nota_biografica">Nota Biográfica</string>
 </resources>
